Restructured devicegen to work more intuitively as a package and enable import statements of the from 'from devicegen import DeviceGenerator'.
