+++ devicegen/__init__.py
@@ -1,0 +1,1 @@
+from devicegen.device_gen import DeviceGenerator
